+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## Version 1.1.0
 - [Responder] Add ViewResponder to return a Response object with twig view
-- [Docs] Add Installation into Readme 
+- [Docs] Add Installation into Readme
+## Version 1.2.0
+- [Responder] Add RedirectResponder to return a RedirectResponder object with route parameters